+++ Changelog.md
@@ -1,5 +1,12 @@
 # Upcoming Changes
 
+# Version 1.24.0
+
+Implement args:
+
+    --align-album-art-top
+    --align-album-art-bottom
+
 # Version 1.23.5
 
 Update bundled dependencies: